backend/requirements.txt
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==2.3.2
Flask-SQLAlchemy==2.5.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==3.1.2
lazy-object-proxy==1.4.0
MarkupSafe==2.0
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
typed-ast==1.4.3
Werkzeug==0.15.6
wrapt==1.11.1
Flask-Cors==3.0.8

# Additional packages installed
# Package            Version
# ------------------ -------
# blinker            1.6.2
# cffi               1.15.1
# click              8.1.6
# cryptography       41.0.2
# ecdsa              0.18.0
# Flask              2.3.2
# Flask-Cors         4.0.0
# Flask-SQLAlchemy   3.0.5
# importlib-metadata 6.8.0
# itsdangerous       2.1.2
# Jinja2             3.1.2
# jose               1.0.0
# MarkupSafe         2.1.3
# pip                22.3.1
# pyasn1             0.5.0
# pycparser          2.21
# PyJWT              2.8.0
# python-jose        3.3.0
# rsa                4.9
# setuptools         65.6.3
# six                1.16.0
# SQLAlchemy         2.0.19
# typing_extensions  4.7.1
# Werkzeug           2.3.6
# zipp               3.16.2